Update logout.php
The original logout script that ships with WoWonder is faulty. Although it clears certain parts of the session like the session id etc, it does not completely destroy the session and therefore the website attempts to perform functions even after you have logged out. This version uses the Php session_destroy() function at the end to make sure that the session is completely gone.